# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HighLevel PayTR Integration API",
    "description": "API documentation for HighLevel PayTR payment integration. This API enables Turkish payment processing through PayTR within the HighLevel CRM platform.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.your-domain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/webhooks/marketplace": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "HighLevel marketplace webhook",
        "description": "Handles app install/uninstall events from HighLevel",
        "operationId": "marketplaceWebhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketplaceWebhook"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "OK"
                }
              }
            }
          },
          "400": {
            "description": "Invalid webhook data"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "LocationHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Location-Id",
        "description": "HighLevel location ID for authentication"
      }
    },
    "schemas": {
...
      "MarketplaceWebhook": {
        "type": "object",
        "required": ["event", "location_id"],
        "properties": {
          "event": {
            "type": "string",
            "enum": ["app.install", "app.uninstall", "app.update"]
          },
          "location_id": {
            "type": "string"
          },
          "company_id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        }
      },
...
    },
...
  }
}